+++ oc2/ochttp.c
@@ -96,13 +96,14 @@ ocfetchurl(CURL* curl, const char* url, OCbytes* buf, long* filetime,
 	int stat = OC_NOERR;
 	CURLcode cstat = CURLE_OK;
 	size_t len;
-    long httpcode = 0;
+        long httpcode = 0;
 	char tbuf[1024];
 	/* Set the URL */
 	cstat = curl_easy_setopt(curl, CURLOPT_URL, (void*)url);
 	if (cstat != CURLE_OK)
 		goto fail;
 	
+#if 0
 	if(creds != NULL && creds->password != NULL  && creds->username != NULL) {
 	    /* Set user and password */
 #if defined (HAVE_CURLOPT_USERNAME) && defined (HAVE_CURLOPT_PASSWORD)
@@ -119,7 +120,7 @@ ocfetchurl(CURL* curl, const char* url, OCbytes* buf, long* filetime,
 			goto fail;
 #endif
 	}
-
+#endif
 	/* send all data to this function  */
 	cstat = curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteMemoryCallback);
 	if (cstat != CURLE_OK)